requirements.txt
Pillow==9.2.0
pytesseract==0.3.9
streamlit==1.10.0
